chore(data-loader): restructure printer functions
