+++ packages/data-loader/src/printers/jsonPrinter.ts
@@ -1,0 +1,7 @@
+import { KintoneRecordField } from "@kintone/rest-api-client";
+
+export const jsonPrinter = (
+  records: Array<{ [k: string]: KintoneRecordField.OneOf }>
+) => {
+  console.log(JSON.stringify(records, null, 2));
+};
